Not the first commit, but the first one to go online. For now, mostly used as a backup from local, but adapted so my code could be run elsewhere too.
Notes, however:
The file SetLocal.R is meant to configure a system with the right Paths. It is in my .gitignore beacuse it's only meant for local use.
A template is public, adapt for your own needs

Input data is not always public, so it's not shared. So nothing to really produce.

Some dumps are way too large for github, I store them locally.
The file OAI Harvest New Records.R rebuilds this dump, but it takes some time.
And it is still a work in progress. For dc-metadata-format it should work, although probalbly not bugfree, but didlmods is not finished.